### 2024-11-22 #7514 @editorialbot
[PRE REVIEW]: scene_synthesizer: A Python Library for Procedural Scene Generation in Robot Manipulation
**Submitting author:** @clemense (<a href="http://orcid.org/0000-0002-5398-4037">Clemens Eppner</a>)
**Repository:** https://github.com/NVlabs/scene_synthesizer
**Branch with paper.md** (empty if default branch):
**Version:** 1.11.4
