### 2024-11-25 #7514 @clemense
Yes, this was brought in from an internal repo. History was squashed.
