### 2024-11-27 #7514 @jromanowska
Hi, sorry to say no, but I don't know anything about the field this software aims at. (I'm also on sick leave this week)
