### 2024-12-03 #7514 @Mechazo11
Hi @crvernon, it will be my pleasure to review this paper. Would be a valuable experience for me. But, before I commit though, what is the tentative deadline to turn in the review?
