### 2024-12-03 #7561 @editorialbot
[REVIEW]: scene_synthesizer: A Python Library for Procedural Scene Generation in Robot Manipulation
**Submitting author:** @clemense (<a href="http://orcid.org/0000-0002-5398-4037">Clemens Eppner</a>)
**Repository:** https://github.com/NVlabs/scene_synthesizer
**Branch with paper.md** (empty if default branch):
**Version:** v1.12.5
