### 2024-12-03 #7561 @crvernon
👋 @clemense, @AlexanderFabisch, and @Mechazo11  - This is the review thread for the paper. All of our communications will happen here from now on.
