### 2024-12-03 NVlabs/scene_synthesizer#1 @AlexanderFabisch
JOSS review: statement of need
Hi @clemense,

I just started the review for JOSS.

The first point on my list that I cannot check is

"A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?"

It is present in the paper, but it should also be part of the documentation.

https://github.com/openjournals/joss-reviews/issues/7561
