### 2024-12-03 NVlabs/scene_synthesizer#2 @AlexanderFabisch
JOSS review: license
I only stumbled across it because the installation instructions say

> This project will download and install additional third-party open source software projects. Review the license terms of these open source projects before use.

The Open USD license is a bit hard to find and it is not a standard license: https://github.com/PixarAnimationStudios/OpenUSD/blob/release/LICENSE.txt

You could directly state which licenses are a bit problematic or maybe just mention / link this license.

https://github.com/openjournals/joss-reviews/issues/7561
