### 2024-12-03 NVlabs/scene_synthesizer#3 @AlexanderFabisch
JOSS review: example scripts - additional dependencies?
When running the example scripts on my computer (after pip install with `[all]`), I get the following errors:

```
$ python examples/stable_poses.py
ValueError: string is not a file: /home/clemens/code/scene_synthesizer/tests/data/assets/shapenetsem_watertight/Mug/b46a9c35587a38f6e87dab59b7e63362.obj

$ python examples/isaaclab_kitchen_env_cfg.py
NameError: name 'configclass' is not defined

$ python examples/add_random_mdl_materials.py
ModuleNotFoundError: No module named 'xatlas'
```

I guess there are some dependencies missing? (+ hardcoded path)

https://github.com/openjournals/joss-reviews/issues/7561
