### 2024-12-03 NVlabs/scene_synthesizer#1 @clemense
I added the statement of need to the documentation: https://scene-synthesizer.github.io/getting_started/statement_of_need.html
