### 2024-12-03 NVlabs/scene_synthesizer#2 @clemense
I added the OpenUSD license link in the latest update (https://github.com/NVlabs/scene_synthesizer/commit/c292bf839d9366ac87c9af4682fb87c23c921582)
