### 2024-12-03 NVlabs/scene_synthesizer#3 @clemense
Thanks!

In the latest update (https://github.com/NVlabs/scene_synthesizer/commit/c292bf839d9366ac87c9af4682fb87c23c921582) I addressed all three issues:

* I removed the hardcoded path from `examples/stable_poses.py`
* I removed the file `example/isaaclab_kitchen_env_cfg.py` since it's not a self-contained runnable example, it was only used as a `literalinclude` in https://scene-synthesizer.github.io/examples/isaaclab_rl.html. I instead added it to the `rst` file directly.
* I added the missing dependency to the `setup.cfg`
