### 2024-12-04 NVlabs/scene_synthesizer#4 @AlexanderFabisch
JOSS review: missing dependencies for unit tests
I installed all dependencies via `pip install -e .[ci]` and run `python -m pytest -n auto tests/`. There are multiple errors, like:

```
E       pxr.Tf.ErrorException:
E               Error in 'pxrInternal_v0_24_11__pxrReserved__::UsdStage::Open' at line 1052 in file /opt/USD/pxr/usd/usd/stage.cpp : 'Failed to open layer @/home/afabisch/Projekte/scene_synthesizer/tests/data/assets/srl_kitchen/scene_overlay.usd@'

...

E               ValueError: /home/afabisch/Projekte/scene_synthesizer/tests/data/assets/partnet_mobility_v0/7236/mobility.urdf is not a file

...

E               ValueError: /home/afabisch/Projekte/scene_synthesizer/tests/data/assets/robots/franka_description/franka_panda.urdf is not a file

...

E               ValueError: /home/afabisch/Projekte/scene_synthesizer/tests/data/assets/partnet_mobility_v0/7236/mobility.urdf is not a file

...

E                   ValueError: string is not a file: /home/afabisch/Projekte/scene_synthesizer/tests/data/assets/shapenetsem/Mug/10f6e09036350e92b3f21f1137c3c347.obj
```

https://github.com/openjournals/joss-reviews/issues/7561#issuecomment-2515256983
