### 2024-12-04 NVlabs/scene_synthesizer#5 @AlexanderFabisch
JOSS review: contribution guidelines
Currently missing:

> Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

More information: https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors

https://github.com/openjournals/joss-reviews/issues/7561
