### 2024-12-05 NVlabs/scene_synthesizer#4 @clemense
Yeah, these assets are not shareable. I assume the preferred way is to remove all the tests?
