### 2024-12-05 NVlabs/scene_synthesizer#4 @AlexanderFabisch
Not shareable as in not public or as in you don't want to add it to the repository? Running a download script before the tests shouldn't be an issue.

Removing all the tests does not make sense. Could you split it into tests with and without internal dependencies and skip the ones with internal dependencies if not available? Could you maybe also replace some of the non-public assets with freely available assets? Which assets do you use in the examples?
