### 2024-12-06 NVlabs/scene_synthesizer#4 @clemense
They are not public and the ones in the examples are procedural ones from the library. With the latest update (https://github.com/NVlabs/scene_synthesizer/commit/51b1c47b31ad13194bb45c68d269a808e3274117) I'm skipping the tests that depend on the internal files.
