### 2024-12-06 NVlabs/scene_synthesizer#4 @AlexanderFabisch
Now I get the following errors:

```
FAILED tests/example_test.py::test_vertical_wall_oven_microwave_cabinet - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/exports_test.py::test_export_primitives_gltf - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/scene_test.py::test_layers - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/scene_test.py::test_collision_manager_with_scene - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/scene_test.py::test_support_as_layer - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/scene_test.py::test_adding_object_transform - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/scene_test.py::test_randomization - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/urdf_export_test.py::test_primitives_urdf_export - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/urdf_export_test.py::test_primitives_urdf_export_single_link - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/urdf_export_test.py::test_primitives_urdf_export_single_link_with_separate_assets - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/scene_test.py::test_part_as_layer - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/urdf_export_test.py::test_primitives_2_urdf_export - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/urdf_export_test.py::test_primitives_2_urdf_export_single_link - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
FAILED tests/scene_test.py::test_container_as_layer - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
ERROR tests/usd_export_test.py
ERROR tests/example_test.py::test_kitchen_collision_bounds - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
ERROR tests/exports_test.py::test_export_urdf_kitchen_collision - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
ERROR tests/example_test.py::test_kitchen_bounds - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
ERROR tests/exports_test.py::test_export_urdf_kitchen_write_meshes - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
ERROR tests/exports_test.py::test_export_urdf_kitchen_collision_write_meshes - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
ERROR tests/exports_test.py::test_export_urdf_kitchen - ImportError: cannot import name 'examples' from 'scene_synthesizer' (/home/...
```
