### 2024-12-06 NVlabs/scene_synthesizer#4 @clemense
Oh, sorry, now these are fixed (https://github.com/NVlabs/scene_synthesizer/commit/5a0450e5205ad5e5142c55e9786a681df2f608ad).
