### 2024-12-06 NVlabs/scene_synthesizer#5 @clemense
Thanks! I added a file `CONTRIBUTION.md` and linked it from the README.  (https://github.com/NVlabs/scene_synthesizer/commit/c4d03f09244b3718d0b2e632e55ecda4d94d50b0)
