### 2024-12-06 NVlabs/scene_synthesizer#4 @AlexanderFabisch
And another dependency is missing here:

```
ImportError while importing test module '/home/afabisch/Projekte/scene_synthesizer/tests/usd_export_test.py'.
Traceback:
../../Data/anaconda3/envs/py311/lib/python3.11/importlib/__init__.py:126: in import_module
tests/usd_export_test.py:15: in <module>
    import meshsets
E   ModuleNotFoundError: No module named 'meshsets'
```
