### 2024-12-06 NVlabs/scene_synthesizer#4 @clemense
Uff, sorry, now it's fully fixed.
