### 2024-12-07 NVlabs/scene_synthesizer#4 @AlexanderFabisch
Great, now it works for me :)

I was also interested in the test coverage. In case you are interested, I changed the pytest configuration a bit in #6

I get 55.2% coverage. Which number do you get with the closed source dependencies?

In any case, I think this issue can be closed.
