### 2024-12-07 #7561 @AlexanderFabisch
I am done with my review. I think the tool is a nice contribution and it's easy to use. I opened several issues in the repository that have all been quickly addressed by @clemense . I have some comments on the paper that you find below my checklist.
