### 2024-12-09 NVlabs/scene_synthesizer#4 @clemense
Great, thank you!  I get 63.0% coverage with those closed source dependencies.
