### 2024-12-16 NVlabs/scene_synthesizer#7 @Mechazo11
JOSS review: Example usage: Adding a ```quick``` example into the primary README.md file
Hello @clemense.
This software looks really useful for RL research, really appreciate the team for making it publicly available.

However, to start playing around with the ```scene synthesizer``` package, one has to navigate all the way into the examples section in the documentation website https://scene-synthesizer.github.io/examples/isaacsim.html

I think if a simpler example could be added after the ```Installation``` section in the ```README.md``` file that walks user through the basic functionalities (i.e. generating a desk with one or two drawers as was discussed here https://scene-synthesizer.github.io/concepts/procedural_assets.html) I think this will make the software more approachable specially to non procedural generation researchers who might just be looking to utilize this software for generating custom training scenarios for their RL experiments.

[openjournals/joss-reviews/issues/7561](https://github.com/openjournals/joss-reviews/issues/7561)
