### 2024-12-16 #7561 @Mechazo11
Hi @crvernon I am done with my review. I left one [ticket](https://github.com/NVlabs/scene_synthesizer/issues/7) in ```scene_synthesizer``` repo regarding adding a simpler example on the ```README.md``` file for quickly portraying the capabilities of the package.
