### 2025-01-04 NVlabs/scene_synthesizer#7 @clemense
Thanks! I just added a small "quick start" section to the README.
