### 2025-01-07 #7561 @clemense
@AlexanderFabisch  Thank you for your comments! Here my answers:

> The difference between this procedural scene generator and others seems to be that it works for physical simulations. It is not clear, however, what makes this use case specifically hard. I'm honestly curious: what are the challenges that this software overcomes?

That's a great question and my answer might not be satisfying: Academic robotics research is largely driven by PhD theses. Although people in general tend to agree that data is the (most) important ingredient in the deep learning era it is impossible to get a PhD by solely focussing on data generation. Instead data generation is a necessary nuisance, and the "intellectual" / scientifcally respected part is the neural network architecture, training process, and resulting application. Case in point: The preprint of this work was rejected by ArXiv (!) due to it not being "scholarly" enough.
scene_synthesizer enables the combination of existing datasets and its own procedural generators to create scene descriptions for robot learning in simulation. I don't know of any tool that is available that does this. As mentioned in the related work section, there are finite datasets or procedural systems that are tightly integrated into bigger frameworks. But IMO they lack the extendability and re-usability that scene_synthesizer offers.

> In addition to this, the summary of features and functionality could be a bit more detailed. The documentation is really awesome and I think the paper could summarize it a bit on a conceptual level.

The paper has a "Features & Functionality" section with high-level descriptions. I'd rather not duplicate the more detailed documentation in the paper since code documentation is better kept in conjuction with code.

> Comparison to similar software packages: I wonder how we would quantify progress in procedural scene generation for physics engines. This seems to be one of the first projects of this kind. What if I wanted to write my own software. How could I claim that it works better? Would we measure time to generate scenes (computational cost)? Is there a way to quantify the quality of the result?

This is a good question, but again very tough to answer. IMO it depends on the downstream task. If the ultimate task is to learn a cat detector vs a bi-pedal running policy, the metrics will be very different. I think that no single metric can capture all of these use cases. What I see a lot in the current wave of LLM-driven scene generators is e.g. using the CLIP score of rendered images as a metric - but it's obvious that this already has a number of shortcomings.

> Example use cases: Again, I am curious how the software was used in these use cases. Could you provide a little bit more context so that I don't have to read all the papers?

The software was used the following ways:
* “Imitating Task and Motion Planning with Visuomotor Transformers”: To generate scenes with tasks for putting objects inside a microwave and shelf. The tasks were then solved with a Task-and-Motion-Planning framework, and the solutions used as data for learning visuo-motor policies.
* “Motion Policy Networks”: To generate cabinet scenes with robot start and goal configurations. These scenarios were used in a motion planner and the resulting collision-free trajectories (together with the scene geometry as a point cloud) was used to train a neural network that imitates the trajectories, thus learning to plan collision-free motions.
* “CabiNet: Scaling Neural Collision Detection for Object Rearrangement with Procedural Scene Generation”: Cabinet and tabletop scenes are used to generate colliding and collision-free configurations. From these examples a network is trained to predict collisions. The network is then used in conjuction with a planner to solve pick-and-place problems.
* “M2T2: Multi-Task Masked Transformer for Object-Centric Pick and Place”: Similar dataset to the previous work.
* “Robo-Point: A Vision-Language Model for Spatial Affordance Prediction for Robotics”: Kitchen scenes with random objects on countertops and inside drawers and cabinets are used to render images and captioning them with spatial descriptions.

> Extensions: It might be interesting for the reader to see how the library could be extended and customized for non-kitchen use cases. This could either be part of the paper or described in the documentation.

That's right, the only kitchen-specific thing in the library are the procedural scenes and some of the procedural assets. The reason there's more kitchen-specific stuff in the library is due to the fact that existing scene generators and existing scene datasets oftentimes avoid kitchens (and focus on other parts of an apartment, bedrooms etc.) due to the highly constraint nature of kitchen furniture and layouts. Currently, the paper doesn't limit itself to kitchen scenes - and the examples in the docs are also not kitchen-specific.

> References: some conferences (CVPR, NeurIPS) are abbreviated. I'd recommend to write the full name because not everybody might know these conferences.

I changed the references and wrote the full name of these conferences.
