### 2025-01-07 #7561 @AlexanderFabisch
> That's a great question and my answer might not be satisfying: Academic robotics research is largely driven by PhD theses. Although people in general tend to agree that data is the (most) important ingredient in the deep learning era it is impossible to get a PhD by solely focussing on data generation. Instead data generation is a necessary nuisance, and the "intellectual" / scientifcally respected part is the neural network architecture, training process, and resulting application. Case in point: The preprint of this work was rejected by ArXiv (!) due to it not being "scholarly" enough.
scene_synthesizer enables the combination of existing datasets and its own procedural generators to create scene descriptions for robot learning in simulation. I don't know of any tool that is available that does this. As mentioned in the related work section, there are finite datasets or procedural systems that are tightly integrated into bigger frameworks. But IMO they lack the extendability and re-usability that scene_synthesizer offers.

Maybe there was a misunderstanding on my side: I thought there were similar software packages that are optimized for visual scenes, but not for physical simulations. I assumed it based on this part of the statement of need: "purely generative models still lack the ability to create scenes that can be used in physics simulator [..]. Other procedural pipelines either focus on learning visual model". My question is how that differs. What differentiates a scene generator for physical simulation from a scene generator for visual "simulation"?

> The preprint of this work was rejected by ArXiv (!) due to it not being "scholarly" enough.

That's crazy. It's more an engineering task, but it is foundational work in today's robotics research. However, to give it a more scientific character, I believe you should focus a bit on how we can evaluate scene generators, so that other people have a way to quantify improvement. That's why I asked for this. Maybe it's too much to ask for in this paper though.

> The software was used the following ways:
>
>     * “Imitating Task and Motion Planning with Visuomotor Transformers”: To generate scenes with tasks for putting objects inside a microwave and shelf. The tasks were then solved with a Task-and-Motion-Planning framework, and the solutions used as data for learning visuo-motor policies.
>
>     * “Motion Policy Networks”: To generate cabinet scenes with robot start and goal configurations. These scenarios were used in a motion planner and the resulting collision-free trajectories (together with the scene geometry as a point cloud) was used to train a neural network that imitates the trajectories, thus learning to plan collision-free motions.
>
>     * “CabiNet: Scaling Neural Collision Detection for Object Rearrangement with Procedural Scene Generation”: Cabinet and tabletop scenes are used to generate colliding and collision-free configurations. From these examples a network is trained to predict collisions. The network is then used in conjuction with a planner to solve pick-and-place problems.
>
>     * “M2T2: Multi-Task Masked Transformer for Object-Centric Pick and Place”: Similar dataset to the previous work.
>
>     * “Robo-Point: A Vision-Language Model for Spatial Affordance Prediction for Robotics”: Kitchen scenes with random objects on countertops and inside drawers and cabinets are used to render images and captioning them with spatial descriptions.

I'd suggest to add that to the paper.

> That's right, the only kitchen-specific thing in the library are the procedural scenes and some of the procedural assets. The reason there's more kitchen-specific stuff in the library is due to the fact that existing scene generators and existing scene datasets oftentimes avoid kitchens (and focus on other parts of an apartment, bedrooms etc.) due to the highly constraint nature of kitchen furniture and layouts. Currently, the paper doesn't limit itself to kitchen scenes - and the examples in the docs are also not kitchen-specific.

... and this as well.
