### 2025-01-08 #7561 @clemense
> Maybe there was a misunderstanding on my side: I thought there were similar software packages that are optimized for visual scenes, but not for physical simulations. I assumed it based on this part of the statement of need: "purely generative models still lack the ability to create scenes that can be used in physics simulator [..]. Other procedural pipelines either focus on learning visual model". My question is how that differs. What differentiates a scene generator for physical simulation from a scene generator for visual "simulation"?

Ah, got it. The physical simulation needs things like collision geometry, mass information, center of mass, friction, restitution, articulation information (joints and their properties, damping, limits, maximum efforts, velocities etc).

> That's crazy. It's more an engineering task, but it is foundational work in today's robotics research. However, to give it a more scientific character, I believe you should focus a bit on how we can evaluate scene generators, so that other people have a way to quantify improvement. That's why I asked for this. Maybe it's too much to ask for in this paper though.

I'm a bit torn on this. Again, it's complicated since scene generation has different objectives, depending on the downstream application/task. The software itself doesn't provide any metrics or support to evaluate scene generators. This is still (and IMO will be for a long time) a fuzzy research area. Also, JOSS explicitly states that "Your paper must not focus on new research results accomplished with the software." and that the software "supports the functioning of research instruments or the execution of research experiments". Keeping the distinction between research software and research results is - IMO - best served by not bloating the paper with random musings about potential metrics and rankings.

> I'd suggest to add that to the paper.

Done!

> I'd suggest to add that to the paper.

Done!
