### 2025-01-08 #7561 @AlexanderFabisch
> Ah, got it. The physical simulation needs things like collision geometry, mass information, center of mass, friction, restitution, articulation information (joints and their properties, damping, limits, maximum efforts, velocities etc).

Could you add that to the paper as well? I think then my review is finished!
