### 2025-01-23 #7561 @crvernon
👋  @clemense - we are almost there! I just need you to address the following:

In your paper:

- [ ] Please remove the backticks from your package name in the title of your paper.  This will cause formatting issues when accepted.

Otherwise, the paper is in great shape.

So here is what we have left to do:

- [ ]  Check the archival deposit (e.g., in Zenodo) to ensure it has the correct metadata. This includes the title (should match the paper title exactly) and author list (should also match the paper exactly). You may also add the authors' ORCID.

- [ ]  Please respond with the DOI of the archived version and the version number you released here.
