### 2025-01-25 #7561 @clemense
Ok, I removed the backticks from the title.
I also did the archival Zenodo release: [10.5281/zenodo.14736777](https://doi.org/10.5281/zenodo.14736777)

Archived version number: v1.12.5
